pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenttester"
version = "0.1.0"
description = "Send prompts to multiple coding agents in parallel and compare results"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
agenttester = "agenttester.cli:app"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["agenttester"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
